telegrambot/requirements.txt
pytelegrambotapi
python-decouple